+++ src/Controller/LogicielController.php
@@ -18,6 +18,8 @@ class LogicielController extends AbstractController
     #[Route('/', name: 'app_logiciel_index', methods: ['GET'])]
     public function index(LogicielRepository $logicielRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_SUPER_ADMIN');
+        
         return $this->render('logiciel/index.html.twig', [
             'logiciels' => $logicielRepository->findAll(),
         ]);
@@ -26,6 +28,8 @@ class LogicielController extends AbstractController
     #[Route('/new', name: 'app_logiciel_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager, ImageUploaderHelper $imageUploaderHelper): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_SUPER_ADMIN');
+        
         $logiciel = new Logiciel();
         $form = $this->createForm(LogicielType::class, $logiciel);
         $form->handleRequest($request);
@@ -56,6 +60,8 @@ class LogicielController extends AbstractController
     #[Route('/{id}', name: 'app_logiciel_show', methods: ['GET'])]
     public function show(Logiciel $logiciel): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_SUPER_ADMIN');
+        
         return $this->render('logiciel/show.html.twig', [
             'logiciel' => $logiciel,
         ]);
@@ -64,6 +70,8 @@ class LogicielController extends AbstractController
     #[Route('/{id}/edit', name: 'app_logiciel_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Logiciel $logiciel, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_SUPER_ADMIN');
+        
         $form = $this->createForm(LogicielType::class, $logiciel);
         $form->handleRequest($request);
 
@@ -82,6 +90,8 @@ class LogicielController extends AbstractController
     #[Route('/{id}', name: 'app_logiciel_delete', methods: ['POST'])]
     public function delete(Request $request, Logiciel $logiciel, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_SUPER_ADMIN');
+        
         if ($this->isCsrfTokenValid('delete'.$logiciel->getId(), $request->getPayload()->get('_token'))) {
             $entityManager->remove($logiciel);
             $entityManager->flush();
